+++ CHANGES.rst
@@ -79,8 +79,8 @@ Unreleased
         allows the user to search for future output of the generator when
         using less and then aborting the program using ctrl-c.
 
--   Parameters can now be deprecated using ``deprecated=True``. These cannot be
-    required nor prompted and are highlighted in usage output. Using them
+-   Parameters can now be deprecated using ``deprecated: bool | str``. These cannot
+    be required nor prompted and are highlighted in usage output. Using them
     will result in a warning being printed. :issue:`2263` :pr:`2271`